requirements.txt
Django==2.1
pytz==2018.5
django-two-factor-auth
